Continue fixing commands code
