+++ lib/cli/cli.js
@@ -2,8 +2,17 @@ import updateNotifier from "update-notifier";
 import yargs from "yargs";
 import {hideBin} from "yargs/helpers";
 import version from "./version.js";
-import add from "./commands/add.js";
-import {fileURLToPath} from "url";
+import base from "./base.js";
+import {fileURLToPath} from "node:url";
+import {readdir} from "node:fs/promises";
+import path from "node:path";
+
+async function getCommands() {
+	const commandsDir = fileURLToPath(new URL("./commands", import.meta.url));
+	return (await readdir(commandsDir, {withFileTypes: true}))
+		.filter((e) => !e.isDirectory() && e.name.endsWith(".js"))
+		.map((e) => path.join(commandsDir, e.name));
+}
 
 export default async (pkg) => {
 	updateNotifier({
@@ -35,11 +44,18 @@ export default async (pkg) => {
 	// Explicitly set script name to prevent windows from displaying "ui5.js"
 	cli.scriptName("ui5");
 
-	// CLI modules
-	/* YError: loading a directory of commands is not supported yet for ESM
-	cli.commandDir("../lib/cli/commands"); */
+	// Setup general options and error handling
+	base(cli);
 
-	cli.command(add);
+	// CLI modules
+	// YError: loading a directory of commands is not supported yet for ESM
+	// cli.commandDir("../lib/cli/commands");
+	// See https://github.com/yargs/yargs/issues/2152
+	const commandModules = await getCommands();
+	for (const modulePath of commandModules) {
+		const {default: command} = await import(modulePath);
+		cli.command(command);
+	}
 
 	// Format terminal output to full available width
 	cli.wrap(cli.terminalWidth());
