+++ lib/cli/commands/init.js
@@ -1,8 +1,10 @@
 // Init
+import baseMiddleware from "../middlewares/base.js";
+
 const initCommand = {
 	command: "init",
 	describe: "Initialize the UI5 Tooling configuration for an application or library project.",
-	middlewares: [require("../middlewares/base.js")]
+	middlewares: [baseMiddleware]
 };
 
 initCommand.handler = async function() {
