+++ lib/cli/commands/remove.js
@@ -1,8 +1,10 @@
 // Remove
+import baseMiddleware from "../middlewares/base.js";
+
 const removeCommand = {
 	command: "remove <framework-libraries..>",
 	describe: "Remove SAPUI5/OpenUI5 framework libraries from the project configuration.",
-	middlewares: [require("../middlewares/base.js")]
+	middlewares: [baseMiddleware]
 };
 
 removeCommand.builder = function(cli) {
