+++ lib/cli/commands/serve.js
@@ -1,11 +1,12 @@
 import path from "path";
 import os from "os";
+import baseMiddleware from "../middlewares/base.js";
 
 // Serve
 const serve = {
 	command: "serve",
 	describe: "Start a web server for the current project",
-	middlewares: [require("../middlewares/base.js")]
+	middlewares: [baseMiddleware]
 };
 
 serve.builder = function(cli) {
