+++ lib/cli/commands/tree.js
@@ -1,10 +1,12 @@
 // Tree
+import baseMiddleware from "../middlewares/base.js";
+
 const tree = {
 	command: "tree",
 	describe:
 		"Outputs the dependency tree of the current project to stdout. " +
 		"It takes all relevant parameters of ui5 build into account.",
-	middlewares: [require("../middlewares/base.js")]
+	middlewares: [baseMiddleware]
 };
 
 tree.builder = function(cli) {
