+++ lib/cli/commands/use.js
@@ -1,8 +1,10 @@
 // Use
+import baseMiddleware from "../middlewares/base.js";
+
 const useCommand = {
 	command: "use <framework-info>",
 	describe: "Initialize or update the project's framework configuration.",
-	middlewares: [require("../middlewares/base.js")]
+	middlewares: [baseMiddleware]
 };
 
 useCommand.builder = function(cli) {
